requirements.txt
pandas
numpy
matplotlib
seaborn
psycopg2
scikit-learn
streamlit
sqlalchemy
python-dotenv